refactor : 장소 업데이트 신청에서 모델 이름을 LocationUpdate에서 UpdateLocation으로 변경
