+++ services/locationService.js
@@ -1,6 +1,6 @@
 const axios = require('axios');
 const Location = require("../models/Location");
-const LocationUpdate = require('../models/LocationUpate');
+const UpdateLocation = require('../models/UpdateLocation');
 const Review = require("../models/Review");
 const CustomError = require("../utils/CustomError");
 const ERROR_CODES = require("../constants/errorCodes");
@@ -54,36 +54,11 @@ exports.updateLocation = async (locationUpdateRequest, userId, locationId) => {
     const location = await Location.findById(locationId);
     if (!location) {
         throw CustomError(ERROR_CODES.NOT_FOUND, ERROR_MESSAGE.LOCATION_NOT_FOUND);
-        return;
     }
     try {
-        const apiUrl = 'https://openapi.naver.com/v1/papago/n2mt';
-        const enNameResponse = await axios.post(apiUrl, {
-            text: locationUpdateRequest.koName,
-            source: "ko",
-            target: "en",
-        }, {
-            headers: {
-                'Content-Type': 'application/x-www-form-urlencoded; charset=UTF-8',
-                'X-Naver-Client-Id': process.env.PAPAGO_CLIENT_ID,
-                'X-Naver-Client-Secret': process.env.PAPAGO_SECRET,
-            },
-        })
-        const enName = enNameResponse.data.message.result.translatedText;
-
-        const enAddressResponse = await axios.post(apiUrl, {
-            text: locationUpdateRequest.koAddress,
-            source: "ko",
-            target: "en",
-        }, {
-            headers: {
-                'Content-Type': 'application/x-www-form-urlencoded; charset=UTF-8',
-                'X-Naver-Client-Id': process.env.PAPAGO_CLIENT_ID,
-                'X-Naver-Client-Secret': process.env.PAPAGO_SECRET,
-            },
-        })
-        const enAddress = enAddressResponse.data.message.result.translatedText;
-        const locationUpdate = await LocationUpdate.create({
+        const enName = await LocationUtil.translateText(locationUpdateRequest.koName, 'ko', 'en');
+        const enAddress = await LocationUtil.translateText(locationUpdateRequest.koAddress, 'ko', 'en');
+        const locationUpdate = await UpdateLocation.create({
             ...locationUpdateRequest,
             user: userId,
             enName,
